pyproject.toml
[project]
name = "harus-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
